feat: enable rrweb recording for hosted link auth flows
- Move activity context manager to context.py to avoid circular imports
- Wrap auth_flow function with activity tracking for hosted_link activities
- Fix circular import between auth_flow.py and mcp/main.py
- Auth browser sessions now recorded and available for replay
